+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - New command: `/leave`, to make the bot leave a discord server.
 - New command: `/versions`, shows the current version and the version of dependencies.
 - New command `/searchworld`, to show filterable list of players online in a server.
+- `/monster` now shows monster's attributes and bestiary info.
 - Watched List now uses an embed, meaning the length is 3 times longer.
 - Minor improvements to documentation site.
 - Improvements to server-log to make them have a uniform style.